fix types in verus dafnybench

+++ benchmarks/verus/dafnybench/files/CVS-Projto1_tmp_tmpb1o0bu8z_proj1_proj1_query.rs
@@ -48,9 +48,9 @@ spec fn mem<T>(x: T, l: List<T>) -> bool
 // </vc-helpers>
 
 // <vc-spec>
-exec fn query(a: &[i32], i: usize, j: usize) -> (s: i32)
+exec fn query(a: &[i8], i: usize, j: usize) -> (s: i8)
     requires 0 <= i <= j <= a.len()
-    ensures s == sum(a@.map(|k, x| x as int), i as int, j as int)
+    ensures s as int == sum(a@.map(|k, x| x as int), i as int, j as int)
 // </vc-spec>
 // <vc-code>
 {
