+++ benchmarks/verus/dafnybench/files/Clover_bubble_sort_BubbleSort.rs
@@ -8,9 +8,9 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn bubble_sort(a: &mut Vec<i32>)
+fn bubble_sort(a: &mut Vec<i8>)
     ensures 
-        forall|i: int, j: int| 0 <= i < j < a.len() ==> a[i] <= a[j],
+        forall|i: int, j: int| 0 <= i < j < a.len() ==> a[i] as int <= a[j] as int,
         a@.to_multiset() == old(a)@.to_multiset(),
 // </vc-spec>
 // <vc-code>
