+++ benchmarks/verus/dafnybench/files/Clover_cal_ans_CalDiv.rs
@@ -8,8 +8,8 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn cal_div() -> (result: (i32, i32))
-  ensures result.0 == 191i32 / 7i32 && result.1 == 191i32 % 7i32,
+fn cal_div() -> (result: (i8, i8))
+  ensures result.0 as i32 == 191i32 / 7i32 && result.1 as i32 == 191i32 % 7i32,
 // </vc-spec>
 // <vc-code>
 {
