+++ benchmarks/verus/dafnybench/files/dafny-synthesis_task_id_262_SplitArray.rs
@@ -8,7 +8,7 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn split_array(arr: &[i32], l: usize) -> (Vec<i32>, Vec<i32>)
+fn split_array(arr: &[i8], l: usize) -> (Vec<i8>, Vec<i8>)
     requires 0 <= l <= arr.len()
 // </vc-spec>
 // <vc-code>
